geolocation-system/ml-models/requirements.txt
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.1.4
scikit-learn==1.3.2
tqdm==4.66.1
